refactor(useTodoManager): ST4-2 — extract trash/filter/deadline-alert into sub-hooks
Extracted 3 self-contained sub-hooks to reduce useTodoManager:
- hooks/useTrashManager.ts (84 lines): trashedTodos, trashedMemos, todoVersions
  state + addTodoVersion, permanentlyDeleteTodo, emptyTrash callbacks
- hooks/useTodoFilters.ts (43 lines): filterState, hasActiveFilters,
  filteredTodos, viewMode
- hooks/useDeadlineAlertSettings.ts (28 lines): deadlineAlertEnabled,
  deadlineAlertMinutes, alertedTodoIdsRef

useTodoManager.ts: 530 → 437 lines. CRUD callbacks kept in parent to avoid
stale-closure risk with mutable parameters. Facade shape unchanged (51 keys).
All 413 tests pass. Mutation on permanentlyDeleteTodo kills 1 test (verified).
